move top8bits as part of bitsplit codec family
